Code improvements & refactor

+++ include/pinout.h
@@ -1,18 +1,13 @@
 #ifndef EVERLASTING_BIRTHDAY_CARD_PINOUT_H
 
 // Microphone pin:
-const int MICROPHONE_PIN = A1;
-
-// RTC pins:
-//const int RTC_CLOCK_PIN = 0;
-//const int RTC_DATA_PIN = 0;
-//const int RTC_RESET_PIN = 0;
+const unsigned short MICROPHONE_PIN = A1;
 
 // Piezo buzzer pin:
-const int BUZZER_PIN = 4;
+const unsigned short BUZZER_PIN = 4;
 
 // Candle LED pin:
-const int CANDLE_PIN = 5;
+const unsigned short CANDLE_PIN = 5;
 
 #define EVERLASTING_BIRTHDAY_CARD_PINOUT_H
